+++ hutool-core/src/main/java/cn/hutool/core/lang/RegexPool.java
@@ -76,7 +76,8 @@ public interface RegexPool {
 	 */
 	String MOBILE_MO = "(?:0|853|\\+853)?(?:|-)6\\d{7}";
 	/**
-	 * 座机号码
+	 * 座机号码<br>
+	 * pr#387@Gitee
 	 */
 	String TEL = "(010|02\\d|0[3-9]\\d{2})-?(\\d{6,8})";
 	/**