+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * 【http   】     HttpRequest.form方法判断集合增强（pr#381@Gitee）
 * 【core   】     NumberUtil增加calculate方法
 * 【core   】     优化TextSimilarity.longestCommonSubstring性能（issue#I42A6V@Gitee）
+* 【core   】     MultipartRequestInputStream改为使用long以支持大文件（issue#I428AN@Gitee）
 
 ### 🐞Bug修复
 * 【core   】     修复RobotUtil双击右键问题（pr#1721@Github）